% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioC_logs.R
\name{bioC_downloads}
\alias{bioC_downloads}
\title{function to download logs from bioConductor stats}
\usage{
bioC_downloads(pckg = NULL, format = "bioC", verbose = TRUE)
}
\arguments{
\item{pckg}{list of packages names}

\item{format}{two options: "bioC" (default) will report the downloads as reported by bioconductor, ie. "Year Month Nb_of_distinct_IPs Nb_of_downloads"; or, "CRAN" will report as CRAN logs does, ie. "Date  Nb_of_downloads package_Name"}

\item{verbose}{boolean flag indicating whether to print information about the processes...}
}
\value{
a list containing a dataframe per package entered with columns as indicated by the format argument
}
\description{
function to download logs from bioConductor stats
}
\examples{
bioC_downloads(c("ABarray","a4Classif"))
bioC_downloads("edgeR",verbose=FALSE)
edgeR.logs <- bioC_downloads("edgeR",format="CRAN")

}
