% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{checkDate}
\alias{checkDate}
\title{function to check date format for "MM-YYYY"}
\usage{
checkDate(date.candidate)
}
\arguments{
\item{date.candidate}{date candidate}
}
\description{
function to check date format for "MM-YYYY"
}
\keyword{internal}
