% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{checkDates}
\alias{checkDates}
\title{function to check dates, ie that t0<t1 and t0!=t1}
\usage{
checkDates(t0, t1)
}
\arguments{
\item{t0}{initial date}

\item{t1}{final date}
}
\value{
a list with t0 being [[1]] and t1 being [[2]]
}
\description{
function to check dates, ie that t0<t1 and t0!=t1
}
\keyword{internal}
