% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{checkValidDates}
\alias{checkValidDates}
\title{function to check dates, ie that t0<t1 and t0 and t1 are in MM-YYYY format}
\usage{
checkValidDates(t0, t1)
}
\arguments{
\item{t0}{initial date}

\item{t1}{final date}
}
\value{
a list with t0 being [[1]] and t1 being [[2]]
}
\description{
function to check dates, ie that t0<t1 and t0 and t1 are in MM-YYYY format
}
\keyword{internal}
