% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{daysInMonth}
\alias{daysInMonth}
\title{function that returns the number of dates in a given date}
\usage{
daysInMonth(d = Sys.Date())
}
\arguments{
\item{d}{string containing a date in YYYY-MM-DD format}
}
\value{
number of days in the given the date
}
\description{
function that returns the number of dates in a given date
}
\keyword{internal}
