% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{last.day.month}
\alias{last.day.month}
\title{function that returns the number of dates in a given month-year}
\usage{
last.day.month(MonthYear)
}
\arguments{
\item{MonthYear}{string containing month and year, ie. MM-YYYY}
}
\value{
number of days --as characters-- in the given month-year
}
\description{
function that returns the number of dates in a given month-year
}
\keyword{internal}
