% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{lst.year}
\alias{lst.year}
\title{function that returns the range dates for the last year}
\usage{
lst.year()
}
\description{
function that returns the range dates for the last year
}
\keyword{internal}
