% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{origin.of.times}
\alias{origin.of.times}
\title{function that provides a "beginning of times" default date}
\usage{
origin.of.times()
}
\description{
function that provides a "beginning of times" default date
}
\keyword{internal}
