% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{year.from.now}
\alias{year.from.now}
\title{function that returns the date from one year ago}
\usage{
year.from.now()
}
\description{
function that returns the date from one year ago
}
\keyword{internal}
