% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs-utils.R
\name{year.to.date}
\alias{year.to.date}
\title{function that returns the range dates for the period year-to-date}
\usage{
year.to.date()
}
\description{
function that returns the range dates for the period year-to-date
}
\keyword{internal}
